Extract listeners logic to plugins
