+++ src/plugins/listener-plugin.ts
@@ -1,0 +1,34 @@
+import Discord, { MessageReaction, Presence, User } from 'discord.js'
+import logger from '../logger'
+
+type ListenerFunction = (...args: any[]) => Promise<void>
+
+const log = logger('listener-plugin')
+
+export default interface ListenerPlugin {
+  discordClient: Discord.Client
+
+  onPresenceUpdate?: (
+    oldPresence: Presence,
+    newPresence: Presence,
+    discordClient: Discord.Client
+  ) => Promise<void>
+
+  onReady?: (discordClient: Discord.Client) => Promise<void>
+
+  onInteractionCreate?: (interaction: Discord.Interaction) => Promise<void>
+
+  onMessageReactionAdd?: (reaction: MessageReaction, user: User) => Promise<void>
+
+  onMessageReactionRemove?: (reaction: MessageReaction, user: User) => Promise<void>
+}
+
+export function wrapErrorHandling(f: ListenerFunction): ListenerFunction {
+  return async (...args) => {
+    try {
+      await f(...args)
+    } catch (error) {
+      log.error('There was an error in the listener:', error)
+    }
+  }
+}
